moved components to seperate files
