+++ app/src/main/java/com/example/tidy/ui/screen/HomeScreen.kt
@@ -18,34 +18,24 @@
 package com.example.tidy.ui.screen
 
 import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.animation.fadeIn
 import androidx.compose.animation.fadeOut
 import androidx.compose.animation.slideInVertically
 import androidx.compose.animation.slideOutVertically
-import androidx.compose.foundation.Canvas
 import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.IntrinsicSize
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.heightIn
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.filled.List
-import androidx.compose.material.icons.filled.CheckBox
-import androidx.compose.material.icons.filled.CheckBoxOutlineBlank
-import androidx.compose.material.icons.filled.ChevronRight
 import androidx.compose.material.icons.filled.Create
 import androidx.compose.material.icons.filled.Delete
 import androidx.compose.material.icons.filled.SkipNext
@@ -58,24 +48,15 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.key
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.rotate
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import androidx.navigation.compose.currentBackStackEntryAsState
 import com.example.tidy.Task
 import com.example.tidy.constants.Routes
-import com.example.tidy.ui.component.taskComponents.TaskCard
+import com.example.tidy.ui.component.subTaskComponents.SubTaskCard
 import com.example.tidy.ui.component.taskComponents.TaskContextAction
-import com.example.tidy.ui.component.taskComponents.TaskDeleteDialog
-import com.example.tidy.ui.component.taskComponents.TaskIconAction
 import com.example.tidy.ui.component.topAppBar.TopAppBar
 import com.example.tidy.viewModels.HomeScreenViewModel
 
@@ -235,172 +216,4 @@ fun contextMenuOptions(
             color = MaterialTheme.colorScheme.error
         )
     )
-}
-
-@Composable
-fun AddIndentation(
-    last: Boolean,
-    map: List<Boolean>
-) { // last is if the task is final child of the parent and map is the map of line and gaps excluding the final one cause a child always attaches to its parent
-    map.forEach { b ->
-        if (b) FullLine() else NoLine()
-        NoLine() // for space of horizontalLine
-    }
-    if (last) HalfLine() else FullLine()
-    HorizontalLine()
-}
-
-@Composable
-fun HorizontalLine() {
-    Box(
-        modifier = Modifier
-            .width(16.dp)
-            .fillMaxHeight(),
-        contentAlignment = Alignment.Center
-    ) {
-        Canvas(modifier = Modifier.fillMaxSize()) {
-            drawLine(
-                color = Color.Gray,
-                start = Offset(0f, size.height / 2),
-                end = Offset(size.width, size.height / 2),
-                strokeWidth = 2f
-            )
-        }
-    }
-}
-
-@Composable
-fun SubTaskCard(
-    task: Task,
-    homeScreenViewModel: HomeScreenViewModel,
-    modifier: Modifier = Modifier,
-    depth: Int = 0,
-    last: Boolean = false,
-    list: List<Boolean> = listOf(),
-) {
-    Column(
-        modifier = modifier
-    ) {
-        var expanded by remember { mutableStateOf(false) }
-        val rotation by animateFloatAsState(
-            targetValue = if (expanded) 90f else 0f,
-            label = "iconRotation"
-        )
-        var showDeleteDialog by remember { mutableStateOf(false) }
-        Row(modifier = Modifier.height(IntrinsicSize.Min)) {
-            if (depth != 0) AddIndentation(last, list)
-            TaskCard(
-                task = task,
-                onClick = {
-                    if (task.children.isNotEmpty()) expanded =
-                        !expanded else homeScreenViewModel.toggleDoneStatus(task)
-                },
-                contextMenuOptions = contextMenuOptions(
-                    task,
-                    homeScreenViewModel
-                ) { showDeleteDialog = true },
-                leadingIcons =
-                    buildList {
-                        if (task.children.isNotEmpty()) {
-                            add(
-                                TaskIconAction(
-                                    icon = Icons.Default.ChevronRight,
-                                    description = "Expand Subtask",
-                                    onClick = { expanded = !expanded },
-                                    modifier = Modifier.rotate(rotation)
-                                )
-                            )
-                        } else {
-                            add(
-                                TaskIconAction(
-                                    icon = if (!task.done) Icons.Default.CheckBoxOutlineBlank else Icons.Default.CheckBox,
-                                    description = "",
-                                    onClick = { homeScreenViewModel.toggleDoneStatus(task) },
-                                )
-                            )
-                        }
-                    },
-            )
-        }
-        if (showDeleteDialog) {
-            TaskDeleteDialog(
-                task = task,
-                onDismiss = { showDeleteDialog = !showDeleteDialog },
-                onDeleteClick = homeScreenViewModel::deleteTask
-            )
-        }
-        if (task.children.isNotEmpty() && expanded) {
-            val bool =
-                task == task.parents.lastOrNull()?.children?.lastOrNull() // is task last child
-            val passingList =
-                if (depth > 0) list + !bool else list // if task is last child then add false no line would be needed
-            task.children.forEach { child ->
-                key(child.id) {
-                    SubTaskCard(
-                        task = child,
-                        homeScreenViewModel = homeScreenViewModel,
-                        depth = depth + 1,
-                        last = child == task.children.last(),
-                        list = passingList
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun FullLine() {
-    Box(
-        modifier = Modifier
-            .width(16.dp)
-            .fillMaxHeight(),
-        contentAlignment = Alignment.Center
-    ) {
-        Canvas(
-            modifier = Modifier
-                .fillMaxHeight()
-                .width(16.dp)
-        ) {
-            drawLine(
-                color = Color.Gray,
-                start = Offset(size.width, 0f),
-                end = Offset(size.width, size.height),
-                strokeWidth = 2f
-            )
-        }
-    }
-}
-
-@Composable
-fun HalfLine() {
-    Box(
-        modifier = Modifier
-            .width(16.dp)
-            .fillMaxHeight(),
-        contentAlignment = Alignment.Center
-    ) {
-        Canvas(
-            modifier = Modifier
-                .fillMaxHeight()
-                .width(16.dp)
-        ) {
-            drawLine(
-                color = Color.Gray,
-                start = Offset(size.width, 0f),
-                end = Offset(size.width, size.height / 2),
-                strokeWidth = 2f
-            )
-        }
-    }
-}
-
-@Composable
-fun NoLine() {
-    Box(
-        modifier = Modifier
-            .width(16.dp)
-            .fillMaxHeight(),
-        contentAlignment = Alignment.Center
-    ) {}
 }
